Separate dimensionality reduction methods
