+++ subcell_pipeline/analysis/dimensionality_reduction/__init__.py
@@ -1,17 +1,1 @@
 """Dimensionality reduction analysis methods and notebooks."""
-
-from .dimensionality_reduction import get_merged_data as get_merged_data
-from .dimensionality_reduction import (
-    plot_fibers_by_key_and_seed as plot_fibers_by_key_and_seed,
-)
-from .dimensionality_reduction import (
-    plot_pacmap_feature_scatter as plot_pacmap_feature_scatter,
-)
-from .dimensionality_reduction import (
-    plot_pca_feature_scatter as plot_pca_feature_scatter,
-)
-from .dimensionality_reduction import (
-    plot_pca_inverse_transform as plot_pca_inverse_transform,
-)
-from .dimensionality_reduction import run_pacmap as run_pacmap
-from .dimensionality_reduction import run_pca as run_pca
