+++ subcell_pipeline/analysis/dimensionality_reduction/_run_pacmap_on_compression_simulations.py
@@ -22,9 +22,11 @@ if __name__ != "__main__":
 # %%
 import pandas as pd
 
-from subcell_pipeline.analysis.dimensionality_reduction import (
+from subcell_pipeline.analysis.dimensionality_reduction.fiber_data import (
     get_merged_data,
     plot_fibers_by_key_and_seed,
+)
+from subcell_pipeline.analysis.dimensionality_reduction.pacmap import (
     plot_pacmap_feature_scatter,
     run_pacmap,
 )
